dbus: Fix compile error for 32-bit platforms
Error encountered with 32-bit gcc 5.3.1:

ell/dbus-kernel.c:412:20: error: cast from pointer to integer of
different size [-Werror=pointer-to-int-cast]

Casting the pointer to uintptr_t instead of uint64_t works on both
32-bit and 64-bit platforms.

+++ ell/dbus-kernel.c
@@ -409,7 +409,7 @@ int _dbus_kernel_send(int fd, size_t bloom_size, uint8_t bloom_n_hash,
 
 	memset(&cmd, 0, sizeof(cmd));
 	cmd.size = sizeof(cmd);
-	cmd.msg_address = (uint64_t) kmsg;
+	cmd.msg_address = (uintptr_t) kmsg;
 
 	ret = ioctl(fd, KDBUS_CMD_SEND, &cmd);
 	if (ret < 0)
